Changed number_density_quasar.py
Note: When using pytest, build (3.8) doesn't apply the "-s" flag to read from stidn while the output is captured. This was caused due to the user being asked how many redshift values they want to generate—changed the code so that only a fixed number of redshift values are to be generated.